Created boilerplate structure
At its core this will be Poetry Python project. For now I named it
'home-finder'. 'GPL-2.0-only' was chosen as this projects licence,
as this is what I usally like to license my stuff with.

I also added a Nix flake as this is what I personally rely on -
dependency handling is nevertheless still handled by Poetry itself
(thanks to pyproject-nix), so things should work portably.

A '.editorconfig' is also supplied - please use it! This allows to
keep a consistent coding style. For more, see https://editorconfig.org.